Name: D. Scott Nettleton Data: 02/27/2013 PURPOSE:  Implemented DLAA antialiasing shader, altered Griswald texture to fix the buckle on his hat. Implemented HUD buttons. FILES:  florinMake         main.cpp         shaderFiles/wsAntialiasing.fsh         textures/btnTemplate.png         textures/chibiGriswaldSkin.png         wsAssets.h         wsAssets/wsButton.cpp         wsAssets/wsButton.h         wsAssets/wsPanel.cpp         wsAssets/wsPanel.h         wsAssets/wsPanelElement.cpp         wsAssets/wsPanelElement.h         wsAssets/wsText.cpp         wsAudio/wsMusic.cpp         wsAudio/wsMusic.h         wsGameFlow/wsController.cpp         wsGameFlow/wsEventManager.h         wsGameFlow/wsGameLoop.cpp         wsGameFlow/wsGameLoop.h         wsGameFlow/wsInputManager.cpp         wsGameFlow/wsInputManager.h         wsGameFlow/wsPointerInput.h         wsGraphics/wsCamera.cpp         wsGraphics/wsRenderSystem.cpp         wsGraphics/wsRenderSystem.h         wsGraphics/wsScreen.cpp         wsGraphics/wsScreen.h         wsGraphics/wsScreenManager.cpp         wsGraphics/wsScreenManager.h         wsGraphics/wsShader.cpp         wsGraphics/wsShader.h         wsUtils/wsOperations.cpp         wsUtils/wsOperations.h

+++ wsUtils/wsOperations.cpp
@@ -26,8 +26,8 @@
 #include <stdlib.h>
 #include <string.h> //  For strlen()
 
-u32 wsScreenWidth = 1280;
-u32 wsScreenHeight = 720;
+f32 wsScreenWidth = 1280;
+f32 wsScreenHeight = 720;
 
 u32 wsCRC32HashFuncTable[256];
 bool wsCRC32HashTableGenerated = false;
